final refactor, unfortunately much similar to the initial
